+++ src/api/services/booksService.ts
@@ -4,11 +4,8 @@ import type {
 } from '@/sources/interfaces';
 import type { BookData } from '@/sources/types';
 import type { AxiosResponse } from 'axios';
-import { AxiosError } from 'axios';
 import { booksApi } from '../axios';
 import { BOOKS_API_KEY } from '@/sources/constants';
-import { messages } from '@/sources/messages';
-import { isApiErrorResponse } from '@/utils/isApiErrorResponse';
 
 export const bookService = {
   getBooksList: async (
@@ -19,71 +16,43 @@ export const bookService = {
     const trimmedQuery = query.trim();
     const startIndex = (page - 1) * pageItemsResults;
 
-    try {
-      const searchQuery = trimmedQuery ? `intitle:${query.trim()}` : 'book';
+    const searchQuery = trimmedQuery ? `intitle:${trimmedQuery}` : 'book';
 
-      const response: AxiosResponse<IBooksListResponse> = await booksApi.get(
-        `?q=${encodeURIComponent(searchQuery)}&startIndex=${startIndex}&maxResults=${pageItemsResults}&langRestrict=en&key=${BOOKS_API_KEY}`
-      );
+    const response: AxiosResponse<IBooksListResponse> = await booksApi.get(
+      `?q=${encodeURIComponent(searchQuery)}&startIndex=${startIndex}&maxResults=${pageItemsResults}&langRestrict=en&key=${BOOKS_API_KEY}`
+    );
 
-      const booksResult: IBookItemResponse[] = Array.isArray(
-        response.data.items
-      )
-        ? response.data.items
-        : [];
+    const booksResult = response.data.items || [];
 
-      const booksList: BookData[] = booksResult
-        .filter(book => book.volumeInfo)
-        .map(book => ({
-          id: book.id,
-          title: book.volumeInfo.title,
-          description: book.volumeInfo.description || '',
-          image:
-            book.volumeInfo.imageLinks?.thumbnail?.replace(
-              /^http:/,
-              'https:'
-            ) || '',
-        }));
+    const booksList: BookData[] = booksResult
+      .filter(book => book.volumeInfo)
+      .map(book => ({
+        id: book.id,
+        title: book.volumeInfo.title,
+        description: book.volumeInfo.description || '',
+        image:
+          book.volumeInfo.imageLinks?.thumbnail?.replace(/^http:/, 'https:') ||
+          '',
+      }));
 
-      const totalItems = response.data.totalItems || 0;
+    const totalItems = response.data.totalItems || 0;
 
-      return { books: booksList, totalItems };
-    } catch (error: unknown) {
-      const message =
-        error instanceof AxiosError && error.response?.data
-          ? isApiErrorResponse(error.response.data)
-            ? error.response.data.error.message
-            : error.message || messages.errorMessage
-          : messages.errorMessage;
-
-      throw new Error(message);
-    }
+    return { books: booksList, totalItems };
   },
 
   getBookById: async (id: string): Promise<BookData> => {
-    try {
-      const response: AxiosResponse<IBookItemResponse> = await booksApi.get(
-        `/${id}?key=${BOOKS_API_KEY}`
-      );
+    const response: AxiosResponse<IBookItemResponse> = await booksApi.get(
+      `/${id}?key=${BOOKS_API_KEY}`
+    );
 
-      const book = response.data;
+    const book = response.data;
 
-      return {
-        id: book.id,
-        description: book.volumeInfo.description || '',
-        authors: book.volumeInfo.authors || '',
-        pageCount: book.volumeInfo.pageCount,
-        printType: book.volumeInfo.printType || '',
-      };
-    } catch (error: unknown) {
-      const message =
-        error instanceof AxiosError && error.response?.data
-          ? isApiErrorResponse(error.response.data)
-            ? error.response.data.error.message
-            : error.message || messages.errorMessage
-          : messages.errorMessage;
-
-      throw new Error(message);
-    }
+    return {
+      id: book.id,
+      description: book.volumeInfo.description || '',
+      authors: book.volumeInfo.authors || '',
+      pageCount: book.volumeInfo.pageCount,
+      printType: book.volumeInfo.printType || '',
+    };
   },
 };
