+++ src/components/button/Button.module.scss
@@ -9,9 +9,9 @@
   border-radius: 5px;
   padding: 3px 40px;
   align-items: center;
-}
 
-button[disabled] {
-  opacity: 0.5;
-  cursor: not-allowed;
+  &:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+  }
 }
