+++ src/pages/book-detail-page/components/BooksDetails.module.scss
