+++ src/pages/main-page/components/books-section/BooksSection.module.scss
