+++ src/pages/main-page/components/books-section/BooksSection.tsx
@@ -1,4 +1,4 @@
-import styles from './ProductsSection.module.scss';
+import styles from './BooksSection.module.scss';
 import { ProductsHeader } from '../../../../components/products-header/ProductsHeader';
 import { titleList } from '@/components/products-header/productsTitleList';
 
@@ -6,7 +6,7 @@ interface Props {
   children: React.ReactNode;
 }
 
-export const ProductsSection = ({ children }: Props) => {
+export const BooksSection = ({ children }: Props) => {
   return (
     <div className={styles.container}>
       <ProductsHeader title={titleList.title} image={titleList.image} />
