+++ src/pages/main-page/components/books-section/components/books-list/BooksList.module.scss
