+++ src/sources/constants.ts
@@ -1,5 +1,6 @@
 const BOOKS_API_KEY = 'AIzaSyAPtAQ26aRgcWR2jK2adRfGoT6zqsYPqdc';
 const BOOKS_API_URL = 'https://www.googleapis.com/books/v1/volumes';
 const RSSCHOOL_URL = 'https://rs.school/courses/reactjs';
+const ITEMS_PER_PAGE = 5;
 
-export { BOOKS_API_KEY, BOOKS_API_URL, RSSCHOOL_URL };
+export { BOOKS_API_KEY, BOOKS_API_URL, RSSCHOOL_URL, ITEMS_PER_PAGE };
